# example/docs/openapi.json
{
  "info": {
    "title": "My API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/echo": {
      "delete": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "",
                  "contentType": "",
                  "method": "DELETE",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Content-Length": "0",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Content-Length": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "example": "bar",
            "in": "query",
            "name": "foo",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "",
                  "contentType": "",
                  "method": "GET",
                  "path": "/echo",
                  "query": {
                    "foo": "bar"
                  },
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {
                        "foo": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "head": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "",
                  "contentType": "",
                  "method": "HEAD",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "options": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "",
                  "contentType": "",
                  "method": "OPTIONS",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Content-Length": "0",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Content-Length": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "example": "application/x-www-form-urlencoded",
            "in": "header",
            "name": "Content-Type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "example": {
                "y": 2
              },
              "schema": {
                "properties": {
                  "y": {
                    "type": "number"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "y=2",
                  "contentType": "application/x-www-form-urlencoded",
                  "method": "PATCH",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Content-Length": "3",
                    "Content-Type": "application/x-www-form-urlencoded",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Content-Length": {
                          "type": "string"
                        },
                        "Content-Type": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "post": {
        "parameters": [
          {
            "example": "application/json",
            "in": "header",
            "name": "Content-Type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "hello": "world"
              },
              "schema": {
                "properties": {
                  "hello": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "{\"hello\":\"world\"}",
                  "contentType": "application/json",
                  "method": "POST",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Content-Length": "17",
                    "Content-Type": "application/json",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Content-Length": {
                          "type": "string"
                        },
                        "Content-Type": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "put": {
        "parameters": [
          {
            "example": "application/x-www-form-urlencoded",
            "in": "header",
            "name": "Content-Type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "example": {
                "x": 1
              },
              "schema": {
                "properties": {
                  "x": {
                    "type": "number"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body": "x=1",
                  "contentType": "application/x-www-form-urlencoded",
                  "method": "PUT",
                  "path": "/echo",
                  "query": {},
                  "requestHeaders": {
                    "Accept": "*/*",
                    "Content-Length": "3",
                    "Content-Type": "application/x-www-form-urlencoded",
                    "Host": "localhost:3000",
                    "User-Agent": "curl/8.7.1"
                  }
                },
                "schema": {
                  "properties": {
                    "body": {
                      "type": "string"
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "query": {
                      "properties": {},
                      "type": "object"
                    },
                    "requestHeaders": {
                      "properties": {
                        "Accept": {
                          "type": "string"
                        },
                        "Content-Length": {
                          "type": "string"
                        },
                        "Content-Type": {
                          "type": "string"
                        },
                        "Host": {
                          "type": "string"
                        },
                        "User-Agent": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "parameters": [
          {
            "example": "bar",
            "in": "query",
            "name": "f",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "Ada",
                    "nullBool": null,
                    "nullFloat": null,
                    "nullInt": null,
                    "nullString": null
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "nullBool": {
                        "type": "null"
                      },
                      "nullFloat": {
                        "type": "null"
                      },
                      "nullInt": {
                        "type": "null"
                      },
                      "nullString": {
                        "type": "null"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "head": {
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "parameters": [
          {
            "example": "Bearer ****",
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "application/json",
            "in": "header",
            "name": "Content-Type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "name": "Charlie"
              },
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "id": 3,
                  "name": "Charlie",
                  "nullBool": null,
                  "nullFloat": null,
                  "nullInt": null,
                  "nullString": null
                },
                "schema": {
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "nullBool": {
                      "type": "null"
                    },
                    "nullFloat": {
                      "type": "null"
                    },
                    "nullInt": {
                      "type": "null"
                    },
                    "nullString": {
                      "type": "null"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Created",
            "headers": {
              "Content-Type": {
                "example": "application/json",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "headers": {
            "Content-Type": {
              "schema": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}